requirements.txt
Flask==2.0.1
requests==2.26.0
google-api-python-client==2.23.0
google-auth==2.2.1
google-cloud-bigquery==2.28.0
google-cloud-firestore==2.3.4
google-cloud-logging==2.6.0


ptvsd==4.3.2 # Required for debugging.
